% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all.interactions.700}
\alias{all.interactions.700}
\title{Cache of protein-protein interactions dataset, as it takes some time to retrieve and process
this will facilitate the vignette building}
\description{
It was filtered with combined_scores and individual scores below 700 without text-based scores
}
\references{
\url{https://string-db.org/}
}
\keyword{data}
