% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loose_rock_functions.R
\name{balanced.cv.folds}
\alias{balanced.cv.folds}
\title{Create balanced folds for cross validation}
\usage{
balanced.cv.folds(..., nfolds = 10)
}
\arguments{
\item{...}{vectors representing data}

\item{nfolds}{number of folds to be created}
}
\value{
list with given input, nfolds and result. The result is a list matching the input with foldid attributed to each position.
}
\description{
Create balanced folds for cross validation
}
\examples{
balanced.cv.folds(1:10, 11:15, nfolds = 2)
balanced.cv.folds(1:10, 11:13, nfolds = 10) # will give a warning
balanced.cv.folds(1:100, 101:133, nfolds = 10)
}
