% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{base.dir}
\alias{base.dir}
\title{change base.dir for run.cache}
\usage{
base.dir(path = NULL)
}
\arguments{
\item{path}{to base directory where cache is saved}
}
\value{
the new path
}
\description{
change base.dir for run.cache
}
\examples{
glmSparseNet:::base.dir("/tmp/cache")
}
