% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{build.function.digest}
\alias{build.function.digest}
\title{Build digest of function from the actual code}
\usage{
build.function.digest(fun)
}
\arguments{
\item{fun}{function call name}
}
\value{
a digest
}
\description{
Build digest of function from the actual code
}
\examples{
glmSparseNet:::build.function.digest(sum)
glmSparseNet:::build.function.digest(c)
}
