% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_db.R
\name{build.string.gene.network}
\alias{build.string.gene.network}
\title{Build gene network from peptide ids}
\usage{
build.string.gene.network(protein.network, use.external.names = FALSE)
}
\arguments{
\item{protein.network}{matrix with colnames and rownames as ensembl peptide id (same order)}

\item{use.external.names}{use external gene names instead of ensembl gene id}
}
\value{
a new matrix with gene ids instead of peptide ids. The size of matrix can be different as
there may not be a mapping or a peptide mapping can have multiple genes.
}
\description{
This can reduce the dimension of the original network, as there may not be a mapping
between peptide and gene id
}
