% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lambda.R
\name{build_lambda}
\alias{build_lambda}
\alias{buildLambda}
\title{Auxiliary function to generate suitable lambda parameters}
\usage{
build_lambda(
  lambda_largest = NULL,
  xdata = NULL,
  ydata = NULL,
  family = NULL,
  orders_of_magnitude_smaller = 3,
  lambda_per_order_magnitude = 150,
  lambda.largest = deprecated(),
  orders.of.magnitude.smaller = deprecated(),
  lambda.per.order.magnitude = deprecated()
)

# deprecated since 1.21.0, please use build_lambda()
buildLambda(
  lambda_largest = NULL,
  xdata = NULL,
  ydata = NULL,
  family = NULL,
  orders_of_magnitude_smaller = 3,
  lambda_per_order_magnitude = 150)
}
\arguments{
\item{xdata}{X parameter for glmnet function}

\item{ydata}{Y parameter for glmnet function}

\item{family}{family parameter to glmnet function}

\item{lambda.largest}{numeric value for largest number of lambda to consider
(usually with a target of 1 selected variable)}

\item{orders.of.magnitude.smaller}{minimum value for lambda
(lambda.largest / 10^orders.of.magnitude.smaller)}

\item{lambda.per.order.magnitude}{how many lambdas to create for each order
of magnitude}
}
\value{
a numeric vector with suitable lambdas
}
\description{
Auxiliary function to generate suitable lambda parameters
}
\examples{
build_lambda(5.4)
}
