% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.index.R
\name{c.index}
\alias{c.index}
\title{Calculates concordance index (c-index) for a given relative risk and survival data}
\usage{
\method{c}{index}(fitted.risk, survival.data, n.cores = 1)
}
\arguments{
\item{fitted.risk}{Vector of fitted risk that can be obtained from a Cox model}

\item{survival.data}{A survival::Surv object or a data object with time and status columns}

\item{n.cores}{number of cores to use}
}
\value{
the c-index
}
\description{
Calculates concordance index (c-index) for a given relative risk and survival data
}
\examples{
# Create the simplest tests data set
library(survival)
test1 <- list(time=c(4,3,1,1,2,2,3),
              status=c(1,1,1,0,1,1,0),
              x=c(0,2,1,1,1,0,0),
              sex=c(0,0,0,0,1,1,1))
# Fit a stratified model
fit1 <- coxph(Surv(time, status) ~ x + strata(sex), test1)
fitted.risk <- predict(fit1, newx = x, type = 'risk')
c.index(fitted.risk, Surv(test1$time, test1$status))
}
