% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{cache.compression}
\alias{cache.compression}
\title{change cache.compression for run.cache}
\usage{
cache.compression(compression = NULL)
}
\arguments{
\item{compression}{see compression parameter in save function}
}
\value{
the new compression
}
\description{
change cache.compression for run.cache
}
\examples{
glmSparseNet:::cache.compression('bzip2')
}
