% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{calculate.result}
\alias{calculate.result}
\title{Calculate/load result and save if necessary}
\usage{
calculate.result(path, compression, force.recalc, show.message, fun, ...)
}
\arguments{
\item{path}{path to save cache}

\item{compression}{compression used in save}

\item{force.recalc}{force to recalculate cache}

\item{show.message}{boolean to show messages}

\item{fun}{function to be called}

\item{...}{arguments to said function
,}
}
\value{
result of fun(...)
}
\description{
This is where the actual work is done
}
\examples{
glmSparseNet:::calculate.result(
  file.path(tempdir(), "calculate.result.Rdata"),
  "gzip",
  FALSE,
  TRUE,
  sum,
  1, 2, 3
)
}
