% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loose_rock_functions.R
\name{coding.genes}
\alias{coding.genes}
\title{Retrive coding genes from known databases}
\usage{
coding.genes(verbose = TRUE)
}
\arguments{
\item{verbose}{show messages with number of genes retrieved}
}
\value{
a table with gene information
}
\description{
It retrieves from NCBI and
}
\examples{
# This can take a few minutes depending on the connection
\dontrun{
    coding.genes()
}
}
