% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{cor.worker}
\alias{cor.worker}
\title{Parallel calculation of correlation}
\usage{
cor.worker(xdata, ix.i, method = "pearson")
}
\description{
Calculates the correlation between a column and all which
 index is above it
}
