% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{create.directory.for.cache}
\alias{create.directory.for.cache}
\title{Create directories for cache}
\usage{
create.directory.for.cache(base.dir, parent.path)
}
\arguments{
\item{base.dir}{tentative base dir to create.}

\item{parent.path}{first 4 characters of digest that will become parent
directory for the actual cache file (this reduces number of files per folder)}
}
\value{
a list of updated base.dir and parent.dir
}
\description{
Create directories for cache
}
\examples{
glmSparseNet:::create.directory.for.cache(tempdir(), 'abcd')
\donttest{
  glmSparseNet:::create.directory.for.cache(
    file.path(getwd(), 'run-cache'), 'abcd'
  )
}
}
