% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmHub.R
\name{cv.glmHub}
\alias{cv.glmHub}
\title{GLMNET cross-validation model penalizing nodes with small degree}
\usage{
cv.glmHub(xdata, ydata, network, network.options = networkOptions(), ...)
}
\arguments{
\item{xdata}{input data, can be a matrix or MultiAssayExperiment.}

\item{ydata}{response data compatible with glmnet.}

\item{network}{type of network, see below.}

\item{network.options}{options to calculate network.}

\item{...}{parameters that \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} accepts.}
}
\value{
an object just as \code{cv.glmnet}
}
\description{
This function overrides the \code{trans.fun} options in \code{network.options} with
an heuristic described in Veríssimo et al. that penalizes nodes with small
degree.
}
\examples{
xdata <- matrix(rnorm(100), ncol = 5)
cv.glmHub(xdata, rnorm(nrow(xdata)), "correlation",
  family = "gaussian",
  nfolds = 5,
  network.options = networkOptions(min.degree = .2)
)
}
\seealso{
Generic function without pre-defined penalization: \code{\link[=cv.glmNetSparse]{cv.glmNetSparse()}}.
Other penalizations: \code{\link[=cv.glmDegree]{cv.glmDegree()}} and \code{\link[=cv.glmOrphan]{cv.glmOrphan()}}.
Model with the same penalization: \code{\link[=glmHub]{glmHub()}}.
}
