% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmOrphan.R
\name{cv.glmOrphan}
\alias{cv.glmOrphan}
\title{GLMNET cross-validation model penalizing nodes with high degree}
\usage{
cv.glmOrphan(
  xdata,
  ydata,
  network,
  options = networkOptions(),
  experiment = NULL,
  network.options = deprecated(),
  experiment.name = deprecated(),
  ...
)
}
\arguments{
\item{xdata}{input data, can be a matrix or MultiAssayExperiment.}

\item{ydata}{response data compatible with glmnet.}

\item{network}{type of network, see below.}

\item{options}{options to calculate network.}

\item{experiment}{name of experiment to use as input in
MultiAssayExperiment object (only if xdata is an object of this class).}

\item{network.options}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{experiment.name}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{...}{parameters that \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} accepts.}
}
\value{
an object just as \code{cv.glmnet}
}
\description{
This function overrides the \code{trans.fun} options in \code{options} with
an heuristic described in Veríssimo et al. that penalizes nodes with high
degree.
}
\examples{
xdata <- matrix(rnorm(100), ncol = 5)
cv.glmOrphan(xdata, rnorm(nrow(xdata)), "correlation",
  family = "gaussian",
  nfolds = 5,
  network.options = networkOptions(min.degree = .2)
)
}
\seealso{
Generic function without pre-defined penalization: \code{\link[=cv.glmNetSparse]{cv.glmNetSparse()}}.
Other penalizations: \code{\link[=cv.glmDegree]{cv.glmDegree()}} and \code{\link[=cv.glmHub]{cv.glmHub()}}.
Model with the same penalization: \code{\link[=glmOrphan]{glmOrphan()}}.
}
