% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.cv.glmnet.R
\docType{methods}
\name{cv.glmSparseNet,SummarizedExperiment-method}
\alias{cv.glmSparseNet,SummarizedExperiment-method}
\title{Calculate GLM model with network-based regularization}
\usage{
\S4method{cv.glmSparseNet}{SummarizedExperiment}(xdata, ydata, network,
  network.options = network.options.default(), ...)
}
\arguments{
\item{xdata}{input data, can be a matrix or MultiAssayExperiment}

\item{ydata}{response data compatible with glmnet}

\item{network}{type of network, see below}

\item{network.options}{options to calculate network}

\item{...}{parameters that cv.glmnet accepts}
}
\value{
an object just as glmnet
}
\description{
Calculate GLM model with network-based regularization
}
