% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{degree.cor}
\alias{degree.cor}
\title{Calculate the degree of the correlation network based on xdata}
\usage{
degree.cor(xdata, cutoff = 0, consider.unweighted = FALSE,
  force.recalc.degree = FALSE, force.recalc.network = FALSE,
  n.cores = parallel:::detectCores(), ...)
}
\arguments{
\item{xdata}{calculate correlation matrix on each column}

\item{cutoff}{positive value that determines a cutoff value}

\item{consider.unweighted}{consider all edges as 1 if they are greater than 0}

\item{n.cores}{number of cores to be used}

\item{...}{extra parameters for fun}

\item{force.recalc}{force recalculation, instead of going to cache}
}
\value{
a vector of the degrees
}
\description{
Calculate the degree of the correlation network based on xdata
}
\examples{
n.col <- 6
xdata <- matrix(rnorm(n.col * 4), ncol = n.col)
degree.cor(xdata)
degree.cor(xdata, cutoff = .5)
degree.cor(xdata, cutoff = .5, consider.unweighted = T)
}
