% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{degree.sparsebn}
\alias{degree.sparsebn}
\title{Calculate degree of correlation matrix}
\usage{
degree.sparsebn(xdata, lambdas.length = 20, cutoff = 0,
  consider.unweighted = TRUE, n.cores = 1, show.message = FALSE,
  force.recalc.degree = FALSE, force.recalc.network = FALSE, ...)
}
\arguments{
\item{xdata}{calculate correlation matrix on each column}

\item{lambdas.length}{sparsebn parameter see sparsebnUtils::sparsebnData documentation}

\item{cutoff}{positive value that determines a cutoff value}

\item{consider.unweighted}{consider all edges as 1 if they are greater than 0}

\item{n.cores}{number of cores to be used}

\item{show.message}{shows cache operation messages}

\item{force.recalc.degree}{force recalculation, instead of going to cache}

\item{force.recalc.network}{force recalculation of network and penalty weights, instead of going to cache}

\item{...}{extra parameters for sparsebn::estimate.dag}
}
\value{
a vector of the degrees
}
\description{
Calculate degree of correlation matrix
}
\examples{
data(cytometryContinuous, package = 'sparsebn')
degree.sparsebn(as.matrix(cytometryContinuous$data))
}
