% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{.baseDir}
\alias{.baseDir}
\title{Change base.dir for run.cache}
\usage{
.baseDir(path = NULL)
}
\arguments{
\item{path}{to base directory where cache is saved}
}
\value{
the new path
}
\description{
Change base.dir for run.cache
}
\examples{
glmSparseNet:::.baseDir("/tmp/cache")
}
