% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.buildFunctionDigest}
\alias{.buildFunctionDigest}
\title{Build digest of function from the actual code}
\usage{
.buildFunctionDigest(fun)
}
\arguments{
\item{fun}{function call name}
}
\value{
a digest
}
\description{
Build digest of function from the actual code
}
\examples{
glmSparseNet:::.buildFunctionDigest(sum)
glmSparseNet:::.buildFunctionDigest(c)
}
