% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{.cacheCompression}
\alias{.cacheCompression}
\title{Change cache.compression for run_cache}
\usage{
.cacheCompression(compression = NULL)
}
\arguments{
\item{compression}{see compression parameter in save function}
}
\value{
the new compression
}
\description{
Change cache.compression for run_cache
}
\examples{
glmSparseNet:::.cacheCompression("bzip2")
}
