% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_apis.R
\name{.curl_workaround}
\alias{.curl_workaround}
\title{Workaround for bug with curl when fetching specific ensembl mirror}
\usage{
.curl_workaround(expr)
}
\arguments{
\item{expr}{expression}
}
\value{
result of expression
}
\description{
Should be solved in issue #39, will test to remove it.
}
\examples{
\donttest{
glmSparseNet:::.curl_workaround({
  biomaRt::useEnsembl(
    biomart = "genes",
    dataset = "hsapiens_gene_ensembl"
  )
})
}
}
