% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.digestCache}
\alias{.digestCache}
\title{Default digest method}
\usage{
.digestCache(val)
}
\arguments{
\item{val}{object to calculate hash over}
}
\value{
a hash of the sha256
}
\description{
Sets a default caching algorithm to use with \code{.runCache}
}
\examples{
glmSparseNet:::.digestCache(c(1, 2, 3, 4, 5))
glmSparseNet:::.digestCache("some example")
}
