% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality-utils.R
\name{.networkGenericParallel}
\alias{.networkGenericParallel}
\title{Calculate the upper triu of the matrix}
\usage{
.networkGenericParallel(
  fun,
  funPrefix,
  xdata,
  buildOutput = "matrix",
  nCores = 1,
  forceRecalcNetwork = FALSE,
  showMessage = FALSE,
  ...
)
}
\arguments{
\item{fun}{function that will calculate the edge weight between 2 nodes}

\item{funPrefix}{used to store low-level information on network as it can
become to large to be stored in memory}

\item{xdata}{base data to calculate network}

\item{buildOutput}{if output returns a 'matrix', 'vector' of the upper triu
without the diagonal or NULL with any other argument}

\item{nCores}{number of cores to be used}

\item{forceRecalcNetwork}{force recalculation, instead of going to cache}

\item{showMessage}{shows cache operation messages}

\item{...}{extra parameters for fun}
}
\value{
depends on buildOutput parameter
}
\description{
Calculate the upper triu of the matrix
}
