% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\docType{data}
\name{.optionsList}
\alias{.optionsList}
\title{Internal list with all the supported options of glmSparseNet with defaults}
\format{
An object of class \code{list} of length 3.
}
\usage{
.optionsList
}
\description{
Internal list with all the supported options of glmSparseNet with defaults
}
\keyword{internal}
