% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.saveRunCache}
\alias{.saveRunCache}
\title{Saving the cache}
\usage{
.saveRunCache(result, path, compression, showMessage)
}
\arguments{
\item{result}{main result to save}

\item{path}{path to the file to save}

\item{compression}{compression method to be used}

\item{show.message}{TRUE to show messages, FALSE otherwise}
}
\value{
result of save operation
}
\description{
Saving the cache
}
\examples{
glmSparseNet:::.saveRunCache(
  35, file.path(tempdir(), "save_run_cache.Rdata"), FALSE, TRUE
)
}
