% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.writeReadme}
\alias{.writeReadme}
\title{Write a file in run-cache directory to explain the origin}
\usage{
.writeReadme(base_dir)
}
\arguments{
\item{base_dir}{directory where to build this file}
}
\value{
the path to the file it has written
}
\description{
Write a file in run-cache directory to explain the origin
}
\examples{

glmSparseNet:::.writeReadme(tempdir())
}
