% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadFileLocal}
\alias{downloadFileLocal}
\title{Download files to local temporary path}
\usage{
downloadFileLocal(urlStr, oD = tempdir())
}
\arguments{
\item{urlStr}{url of file to download}

\item{oD}{temporary directory to store file}
}
\value{
path to file
}
\description{
In case of new call it uses the temporary cache instead of
downloading again.
}
\details{
Inspired by STRINGdb Bioconductor package, but using curl
as file may be too big to handle.
}
\examples{
glmSparseNet:::downloadFileLocal(
  'https://string-db.org/api/tsv-no-header/version')
}
