% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.kaplan.R
\docType{methods}
\name{draw.kaplan,list,data.frame,data.frame-method}
\alias{draw.kaplan,list,data.frame,data.frame-method}
\title{draw.kaplan}
\usage{
\S4method{draw.kaplan}{list,data.frame,data.frame}(chosen.btas, xdata,
  ydata, probs = c(0.5, 0.5), plot.title = "SurvivalCurves",
  xlim = NULL, ylim = NULL, expand.yzero = FALSE,
  legend.outside = FALSE)
}
\arguments{
\item{chosen.btas}{list.}

\item{xdata}{data.frame.}

\item{ydata}{data.frame.}

\item{probs}{How to separate high and low risk patients 50\%-50\% is the default, but for top and bottom 40\% -> c(.4,.6)}

\item{plot.title}{Name of file if}

\item{xlim}{Optional argument to limit the x-axis view}

\item{ylim}{Optional argument to limit the y-axis view}

\item{expand.yzero}{expand to y = 0}

\item{legend.outside}{If TRUE legend will be outside plot, otherwise inside}
}
\value{
object with logrank test and kaplan-meier survival plot
}
\description{
draw.kaplan
}
