% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.kaplan.R
\name{draw.kaplan}
\alias{draw.kaplan}
\title{draw.kaplan deprecation warning}
\usage{
draw.kaplan(...)
}
\arguments{
\item{...}{see separate2groups.cox}
}
\value{
object with logrank test and kaplan-meier survival plot
}
\description{
draw.kaplan deprecation warning
}
\seealso{
separate2groups.cox
}
