% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_apis.R
\name{ensembl.gene.names}
\alias{ensembl.gene.names}
\title{Retrieve ensembl gene names from biomaRt}
\usage{
ensembl.gene.names(gene.id)
}
\arguments{
\item{gene.id}{character vector with gene names}
}
\value{
a dataframe with external gene names, ensembl_id
}
\description{
Retrieve ensembl gene names from biomaRt
}
\examples{
ensembl.gene.names(c('MOB1A','RFLNB', 'SPIC', 'TP53'))
}
