% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_apis.R
\name{geneNames}
\alias{geneNames}
\title{Retrieve gene names from biomaRt}
\usage{
geneNames(ensembl.genes, use.cache = TRUE)
}
\arguments{
\item{ensembl.genes}{character vector with gene names in ensembl_id format}

\item{use.cache}{Boolean indicating if biomaRt cache should be used}
}
\value{
a dataframe with external gene names, ensembl_id
}
\description{
Retrieve gene names from biomaRt
}
\examples{
geneNames(c('ENSG00000114978','ENSG00000166211', 'ENSG00000183688'))
}
