% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmDegree.R
\name{glmDegree}
\alias{glmDegree}
\title{GLMNET model penalizing nodes with small degree}
\usage{
glmDegree(xdata, ydata, network, network.options = networkOptions(), ...)
}
\arguments{
\item{xdata}{input data, can be a matrix or MultiAssayExperiment.}

\item{ydata}{response data compatible with glmnet.}

\item{network}{type of network, see below.}

\item{network.options}{options to calculate network.}

\item{...}{parameters that \code{\link[glmnet:glmnet]{glmnet::glmnet()}} accepts.}
}
\value{
an object just as glmnet
}
\description{
This function overrides the \code{trans.fun} options in \code{network.options} with the
inverse of a degree described in Veríssimo et al. (2015) that penalizes
nodes with small degree.
}
\examples{
xdata <- matrix(rnorm(100), ncol = 5)
glmDegree(xdata, rnorm(nrow(xdata)), "correlation",
  family = "gaussian",
  network.options = networkOptions(min.degree = .2)
)
}
\seealso{
Generic function without pre-defined penalization: \code{\link[=glmNetSparse]{glmNetSparse()}}.
Other penalizations: \code{\link[=glmHub]{glmHub()}} and \code{\link[=glmOrphan]{glmOrphan()}}.
Cross-validation with the same penalization: \code{\link[=cv.glmDegree]{cv.glmDegree()}}.
}
