% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.glmnet.R
\docType{methods}
\name{glmSparseNet,MultiAssayExperiment-method}
\alias{glmSparseNet,MultiAssayExperiment-method}
\title{Calculate GLM model with network-based regularization}
\usage{
\S4method{glmSparseNet}{MultiAssayExperiment}(xdata, ydata, network,
  experiment.name = NULL, network.options = network.options.default(), ...)
}
\arguments{
\item{xdata}{input data, can be a matrix or MultiAssayExperiment}

\item{ydata}{response data compatible with glmnet}

\item{network}{type of network, see below}

\item{experiment.name}{name of experiment to use as input in MultiAssayExperiment object}

\item{network.options}{options to calculate network}

\item{...}{parameters that glmnet accepts}
}
\value{
an object just as glmnet
}
\description{
Calculate GLM model with network-based regularization
}
\details{
* string to calculate network based on data (correlation, covariance)
 * matrix representing the network
 * vector with already calculated penalty weights (can also be used directly with glmnet)
}
