% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSparseNet-package.R
\docType{package}
\name{glmSparseNet-package}
\alias{glmSparseNet-package}
\title{glmSparseNet: Network Centrality Metrics for Elastic-Net Regularized Models}
\description{
glmSparseNet is an R-package that generalizes sparse regression models when the features (e.g. genes) have a graph structure (e.g. protein-protein interactions), by including network-based regularizers. glmSparseNet uses the glmnet R-package, by including centrality measures of the network as penalty weights in the regularization. The current version implements regularization based on node degree, i.e. the strength and/or number of its associated edges, either by promoting hubs in the solution or orphan genes in the solution. All the glmnet distribution families are supported, namely "gaussian", "poisson", "binomial", "multinomial", "cox", and "mgaussian".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.github.com/sysbiomed/glmSparseNet}
  \item Report bugs at \url{https://www.github.com/sysbiomed/glmSparseNet/issues}
}

}
\author{
\strong{Maintainer}: André Veríssimo \email{andre.verissimo@tecnico.ulisboa.pt} (\href{https://orcid.org/0000-0002-2212-339X}{ORCID})

Authors:
\itemize{
  \item Susana Vinga \email{susanavinga@tecnico.ulisboa.pt}
}

Other contributors:
\itemize{
  \item Eunice Carrasquinha \email{eunice.trigueirao@tecnico.ulisboa.pt} [contributor]
  \item Marta Lopes \email{marta.lopes@tecnico.ulisboa.pt} [contributor]
}

}
\keyword{internal}
