% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loose_rock_functions.R
\name{glmSparseNet.options}
\alias{glmSparseNet.options}
\title{Constants for 'glmSparseNet' package}
\usage{
glmSparseNet.options(..., simplify = FALSE, update = list())
}
\arguments{
\item{...}{TODO}

\item{simplify}{TODO}

\item{update}{pair list of update to options}
}
\description{
Log level constants and the logger options.
}
\details{
The logging configuration is managed by 'glmSparseNet.options', a function
generated by OptionsManager within 'futile.options'.
}
\seealso{
\code{futile.options}
}
