% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.generic.R
\docType{methods}
\name{glmSparseNet.private,ANY,MultiAssayExperiment-method}
\alias{glmSparseNet.private,ANY,MultiAssayExperiment-method}
\title{Calculate GLM model with network-based regularization}
\usage{
\S4method{glmSparseNet.private}{ANY,MultiAssayExperiment}(fun, xdata,
  ydata, network, experiment.name = NULL,
  network.options = network.options.default(), ...)
}
\arguments{
\item{fun}{function to be called (glmnet or cv.glmnet)}

\item{xdata}{input data, can be a matrix or MultiAssayExperiment}

\item{ydata}{response data compatible with glmnet}

\item{network}{type of network, see below}

\item{network.options}{options to calculate network}

\item{...}{parameters that glmnet accepts}
}
\value{
an object just as glmnet
}
\description{
Calculate GLM model with network-based regularization
}
