% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_apis.R
\name{hallmarks}
\alias{hallmarks}
\title{Retrieve hallmarks of cancer count for genes}
\usage{
hallmarks(genes, measure = "count", hallmarks = "full")
}
\arguments{
\item{genes}{gene names}

\item{measure}{see below}

\item{hallmarks}{see below}
}
\value{
data.frame with choosen measure and hallmarks.
It also returns a vector with genes that do not have any
hallmarks

See http://chat.lionproject.net/api for more details on the
measure and hallmarks parameters
}
\description{
Retrieve hallmarks of cancer count for genes
}
\examples{
hallmarks(c('MOB1A', 'RFLNB', 'SPIC'))
}
