% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.options.R
\name{heuristicScale}
\alias{heuristicScale}
\title{Heuristic function to use in high dimensions}
\usage{
heuristicScale(
  x,
  subExp10 = -1,
  expMult = -1,
  subExp = -1,
  sub.exp10 = deprecated(),
  exp.mult = deprecated(),
  sub.exp = deprecated()
)
}
\arguments{
\item{x}{vector of values to scale}

\item{sub.exp10}{value to subtract to base 10 exponential, for example:
\code{10^0 - sub.exp10 = 1 - sub.exp10}}

\item{exp.mult}{parameter to multiply exponential, i.e. to have a negative
exponential or positive}

\item{sub.exp}{value to subtract for exponentional, for example if x = 0,
\code{exp(0) - sub.exp = 1 - sub.exp}}
}
\value{
a vector of scaled values
}
\description{
Heuristic function to use in high dimensions
}
\examples{
heuristicScale(rnorm(1:10))
}
