% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.options.R
\name{hub.heuristic}
\alias{hub.heuristic}
\title{Heuristic function to penalize nodes with low degree}
\usage{
hub.heuristic(x)
}
\arguments{
\item{x}{single value of vector}
}
\value{
transformed
}
\description{
Heuristic function to penalize nodes with low degree
}
\examples{
hub.heuristic(rnorm(1:10))
}
