% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_options.R
\name{hubHeuristic}
\alias{hubHeuristic}
\title{Heuristic function to penalize nodes with low degree}
\usage{
hubHeuristic(x)
}
\arguments{
\item{x}{single value of vector}
}
\value{
transformed
}
\description{
Heuristic function to penalize nodes with low degree
}
\examples{
hubHeuristic(rnorm(1:10))
}
