% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.index.R
\name{my.c.index.cmp}
\alias{my.c.index.cmp}
\title{Concordance index comparison between a pair}
\usage{
my.c.index.cmp(risk.a, risk.b, time.a, time.b, event.a, event.b)
}
\arguments{
\item{risk.a}{risk of individual A}

\item{risk.b}{risk of indiviudal B}

\item{time.a}{time of event of individual A}

\item{time.b}{time of event of individual B}

\item{event.a}{has event occurred in individual A}

\item{event.b}{has event occurred in individual B}
}
\value{
the pairwise comparison
}
\description{
Concordance index comparison between a pair
}
\examples{
my.c.index.cmp(1, 2, 10, 30, TRUE, FALSE)
my.c.index.cmp(1, 2, 10, 5, TRUE, FALSE)
my.c.index.cmp(1, 2, 10, 5, TRUE, TRUE)
}
