% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_symbol.R
\name{my.colors}
\alias{my.colors}
\title{Custom pallete of colors}
\usage{
my.colors(ix = NULL)
}
\arguments{
\item{ix}{index for a color}
}
\value{
a color
}
\description{
Custom pallete of colors
}
\examples{
my.colors()
my.colors(5)
}
