% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_symbol.R
\name{my.symbols}
\alias{my.symbols}
\title{Custom pallete of symbols in plots}
\usage{
my.symbols(ix = NULL)
}
\arguments{
\item{ix}{index for symbol}
}
\value{
a symbol
}
\description{
Custom pallete of symbols in plots
}
\examples{
my.symbols()
my.symbols(2)
}
