% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_symbol.R, R/deprecation-utils.R
\name{myColors}
\alias{myColors}
\alias{my.colors}
\title{Custom pallete of colors}
\usage{
myColors(ix = NULL)

# deprecated, please use myColors()
my.colors(ix = NULL)
}
\arguments{
\item{ix}{index for a color}
}
\value{
a color
}
\description{
Custom pallete of colors
}
\examples{
myColors()
myColors(5)
}
