% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_symbol.R, R/deprecation-utils.R
\name{mySymbols}
\alias{mySymbols}
\alias{my.symbols}
\title{Custom pallete of symbols in plots}
\usage{
mySymbols(ix = NULL)

# deprecated, please use mySymbols()
my.symbols(ix = NULL)
}
\arguments{
\item{ix}{index for symbol}
}
\value{
a symbol
}
\description{
Custom pallete of symbols in plots
}
\examples{
mySymbols()
mySymbols(2)
}
