% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.glmnet.R
\docType{methods}
\name{network.glmnet,MultiAssayExperiment-method}
\alias{network.glmnet,MultiAssayExperiment-method}
\title{Calculate GLM model with network-based regularization}
\usage{
\S4method{network.glmnet}{MultiAssayExperiment}(xdata, ydata, network,
  experiment.name = NULL, network.options = network.options.default(), ...)
}
\arguments{
\item{xdata}{input data as a MultiAssayExperiment}

\item{ydata}{response data compatible with glmnet}

\item{network}{type of network, see below}

\item{experiment.name}{name of experiment to use as input in MultiAssayExperiment object}

\item{network.options}{options to calculate network}

\item{...}{parameters that glmnet accepts}
}
\value{
an object just as glmnet
}
\description{
Calculate GLM model with network-based regularization
}
