% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.glmnet.R
\docType{methods}
\name{network.glmnet,SummarizedExperiment-method}
\alias{network.glmnet,SummarizedExperiment-method}
\title{Calculate GLM model with network-based regularization}
\usage{
\S4method{network.glmnet}{SummarizedExperiment}(xdata, ydata, network,
  network.options = network.options.default(), ...)
}
\arguments{
\item{xdata}{input data as a SummarizedExperiment}

\item{ydata}{response data compatible with glmnet}

\item{network}{type of network, see below}

\item{network.options}{options to calculate network}

\item{...}{parameters that glmnet accepts}
}
\value{
an object just as glmnet
}
\description{
Calculate GLM model with network-based regularization
}
