% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.generic.R
\docType{methods}
\name{network.glmnet.private,ANY,matrix-method}
\alias{network.glmnet.private,ANY,matrix-method}
\title{Calculate GLM model with network-based regularization}
\usage{
\S4method{network.glmnet.private}{ANY,matrix}(fun, xdata, ydata, network,
  network.options = network.options.default(), ...)
}
\arguments{
\item{xdata}{matrix.}
}
\value{
an object just as glmnet
}
\description{
Calculate GLM model with network-based regularization
}
