% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.generic.R
\docType{methods}
\name{network.glmnet.private,ANY,MultiAssayExperiment-method}
\alias{network.glmnet.private,ANY,MultiAssayExperiment-method}
\title{Calculate GLM model with network-based regularization}
\usage{
\S4method{network.glmnet.private}{ANY,MultiAssayExperiment}(fun, xdata, ydata,
  network, experiment.name = NULL,
  network.options = network.options.default(), ...)
}
\arguments{
\item{xdata}{MultiAssayExperiment.}
}
\value{
an object just as glmnet
}
\description{
Calculate GLM model with network-based regularization
}
