% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.generic.R
\name{network.glmnet.private}
\alias{network.glmnet.private}
\title{Calculate GLM model with network-based regularization}
\usage{
network.glmnet.private(fun, xdata, ydata, network,
  network.options = network.options.default(), ...)
}
\arguments{
\item{fun}{function to be called (glmnet or cv.glmnet)}

\item{xdata}{input data, can be a matrix or MultiAssayExperiment}

\item{ydata}{response data compatible with glmnet}

\item{network}{type of network, see below}

\item{network.options}{options to calculate network}

\item{...}{parameters that glmnet accepts}
}
\value{
an object just as glmnet

network parameter accepts:

 * string to calculate network based on data (correlation, covariance)
 * matrix representing the network
 * vector with already calculated penalty weights (can also be used directly with glmnet)
}
\description{
Calculate GLM model with network-based regularization
}
