% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.options.R
\name{network.options.default}
\alias{network.options.default}
\title{Setup network options}
\usage{
network.options.default(method = "pearson", unweighted = TRUE, cutoff = 0,
  centrality = "degree", min.degree = 0, n.cores = 1,
  trans.fun = function(x) {     x })
}
\arguments{
\item{method}{in case of correlation and covariance, which method to use}

\item{unweighted}{calculate degree using unweighted network}

\item{cutoff}{cuttoff value in network edges to trim the network}

\item{centrality}{centrality measure to use, currently only supports degree}

\item{n.cores}{number of cores to use, default to 1}
}
\value{
a list of options
}
\description{
Setup network options, such as using weighted or unweighted degree,
 which centrality measure to use
}
\examples{
network.options.default(unweighted = FALSE)
}
