% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{network.worker}
\alias{network.worker}
\title{Worker to calculate edge weight for each pair of ix.i node and following}
\usage{
network.worker(fun, xdata, ix.i, ...)
}
\arguments{
\item{fun}{function to be used, can be cor, cov or any other defined function}

\item{xdata}{original data to calculate the function over}

\item{ix.i}{starting index, this can be used to save ony upper triu}

\item{...}{extra parameters for fun}
}
\value{
a vector with size `ncol(xdata) - ix.i`
}
\description{
Note that it assumes it does not calculate for index below and equal to ix.i
}
\examples{
network.worker(cor, matrix(rnorm(20*10), ncol = 10), 1)
network.worker(cor, matrix(rnorm(20*10), ncol = 10), 5)
}
