% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.options.R
\name{orphan.heuristic}
\alias{orphan.heuristic}
\title{Heuristic function to penalize nodes with high degree}
\usage{
orphan.heuristic(x)
}
\arguments{
\item{x}{single value of vector}
}
\value{
transformed
}
\description{
Heuristic function to penalize nodes with high degree
}
\examples{
orphan.heuristic(rnorm(1:10))
}
