% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data_tcga.R
\name{prepare.tcga.survival.data}
\alias{prepare.tcga.survival.data}
\title{Prepare FPKM data from TCGA project (specific tissue)}
\usage{
prepare.tcga.survival.data(project = "brca",
  tissue.type = "primary.solid.tumor", handle.duplicates = "keep_first",
  coding.genes = FALSE)
}
\arguments{
\item{tissue.type}{}
}
\description{
This function will load data and pre-process it to be used in
 survival models.
}
\details{
It will:
 * load data
 *
}
\examples{
prepare.tcga.survival.data('brca', 'primary.solid.tumor', 'keep_first')
}
