% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loose_rock_functions.R
\name{proper}
\alias{proper}
\title{Capitalizes all words in string}
\usage{
proper(x)
}
\arguments{
\item{x}{String}
}
\value{
a capitalized string (all words)
}
\description{
Capitalizes all words in string
}
\examples{
proper('i saw a dEaD parrot')
}
