% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_apis.R
\name{protein.to.ensembl.gene.names}
\alias{protein.to.ensembl.gene.names}
\title{Retrieve ensembl gene ids from proteins}
\usage{
protein.to.ensembl.gene.names(ensembl.proteins)
}
\arguments{
\item{ensembl.proteins}{character vector with gene names in ensembl_peptide_id format}
}
\value{
a dataframe with external gene names, ensembl_peptide_id
}
\description{
Retrieve ensembl gene ids from proteins
}
\examples{
protein.to.ensembl.gene.names(c('ENSP00000235382','ENSP00000233944', 'ENSP00000216911'))
}
