% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.options.R
\name{reduceByExperiment}
\alias{reduceByExperiment}
\title{Filter MultiAssayExperiment colData for specific experiment}
\usage{
reduceByExperiment(multi.assay, experiment.name)
}
\arguments{
\item{multi.assay}{MultiAssayExperiment object}

\item{experiment.name}{name of experiment}
}
\value{
a MultiExperimentAssay with only clinical data of that experiment
}
\description{
Filter MultiAssayExperiment colData for specific experiment
}
\examples{
multi.assay <- MultiAssayExperiment::miniACC
reduceByExperiment(multi.assay, 'RNASeq2GeneNorm')
}
