% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.PRAD.survival}
\alias{sample.PRAD.survival}
\title{TCGA Prostate cancer sample data set to use with survival model in vignettes}
\description{
TCGA Prostate cancer sample data set to use with survival model in vignettes
}
\references{
\url{https://portal.gdc.cancer.gov/}
}
\keyword{data}
