% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.SKCM.survival}
\alias{sample.SKCM.survival}
\title{TCGA Skin melanoma cancer sample data set to use with survival model in vignettes}
\description{
TCGA Skin melanoma cancer sample data set to use with survival model in vignettes
}
\references{
\url{https://portal.gdc.cancer.gov/}
}
\keyword{data}
