% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{save.run.cache}
\alias{save.run.cache}
\title{Saving the cache}
\usage{
save.run.cache(result, path, compression, show.message)
}
\arguments{
\item{result}{main result to save}

\item{path}{path to the file to save}

\item{compression}{compression method to be used}

\item{show.message}{TRUE to show messages, FALSE otherwise}
}
\value{
result of save operation
}
\description{
Saving the cache
}
\examples{
glmSparseNet:::save.run.cache(
  35, file.path(tempdir(), "save.run.cache.Rdata"), FALSE, TRUE
)
}
