% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{show.message}
\alias{show.message}
\title{Show messages option in run.cache}
\usage{
show.message(show.message = NULL)
}
\arguments{
\item{show.message}{boolean indicating to show messages or not}
}
\value{
the show.message option
}
\description{
Show messages option in run.cache
}
\examples{
glmSparseNet:::show.message(FALSE)
}
