% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{string.network.700.cache}
\alias{string.network.700.cache}
\title{Cache of protein-protein network, as it takes some time to
retrieve and process this will facilitate the vignette building}
\format{
An object of class \code{dgCMatrix} with 11033 rows and 11033 columns.
}
\usage{
string.network.700.cache
}
\description{
It was filtered with combined_scores and individual scores below 700 without
text-based scores
}
\examples{
head(string.network.700.cache)
}
\references{
\url{https://string-db.org/}
}
\keyword{data}
