% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_db.R
\name{stringDBhomoSapiens}
\alias{stringDBhomoSapiens}
\title{Download protein-protein interactions from STRING DB}
\usage{
stringDBhomoSapiens(
  version = "11.0",
  scoreThreshold = 0,
  removeText = TRUE,
  score_threshold = deprecated(),
  remove.text = deprecated()
)
}
\arguments{
\item{version}{version of the database to use}

\item{scoreThreshold}{remove scores below threshold}

\item{removeText}{remove text mining-based scores}

\item{score_threshold}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{remove.text}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\value{
a data.frame with rows representing an interaction between two
proteins, and columns
the count of scores above the given score_threshold
}
\description{
Download protein-protein interactions from STRING DB
}
\examples{
\donttest{
stringDBhomoSapiens(scoreThreshold = 800)
}
}
