% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{tempdir.cache}
\alias{tempdir.cache}
\title{Temporary directory for runCache}
\usage{
tempdir.cache()
}
\value{
a path to a temporary directory used by runCache
}
\description{
Temporary directory for runCache
}
