% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write.readme}
\alias{write.readme}
\title{Write a file in run-cache directory to explain the origin}
\usage{
write.readme(base.dir)
}
\arguments{
\item{base.dir}{directory where to build this file}
}
\description{
Write a file in run-cache directory to explain the origin
}
\examples{
glmSparseNet:::write.readme(tempdir())
}
