% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_data.R
\name{balanced.cv.folds}
\alias{balanced.cv.folds}
\title{Create balanced folds for cross validation}
\usage{
balanced.cv.folds(..., nfolds = 10)
}
\arguments{
\item{...}{vectors representing data}

\item{nfolds}{number of folds to be created}
}
\value{
list with given input, nfolds and result. The result is a list
matching the input with foldid attributed to each position.
}
\description{
Create balanced folds for cross validation
}
\examples{
glmSparseNet:::balanced.cv.folds(seq(10), seq(11, 15), nfolds = 2)
# will give a warning
glmSparseNet:::balanced.cv.folds(seq(10), seq(11, 13), nfolds = 10) 
glmSparseNet:::balanced.cv.folds(seq(100), seq(101, 133), nfolds = 10)
}
