% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{buildLambda}
\alias{buildLambda}
\title{Auxiliary function to generate suitable lambda parameters}
\usage{
buildLambda(
  lambda.largest = NULL,
  xdata = NULL,
  ydata = NULL,
  family = NULL,
  orders.of.magnitude.smaller = 3,
  lambda.per.order.magnitude = 150
)
}
\arguments{
\item{lambda.largest}{numeric value for largest number of lambda to consider
(usually with a target of 1 selected variable)}

\item{xdata}{X parameter for glmnet function}

\item{ydata}{Y parameter for glmnet function}

\item{family}{family parameter to glmnet function}

\item{orders.of.magnitude.smaller}{minimum value for lambda
(lambda.largest / 10^orders.of.magnitude.smaller)}

\item{lambda.per.order.magnitude}{how many lambdas to create for each order
of magnitude}
}
\value{
a numeric vector with suitable lambdas
}
\description{
Auxiliary function to generate suitable lambda parameters
}
\examples{
buildLambda(5.4)
}
