% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.options.R
\name{calc.penalty}
\alias{calc.penalty}
\title{Calculate penalty based on data}
\usage{
calc.penalty(xdata, penalty.type, network.options = network.options.default())
}
\arguments{
\item{xdata}{input data}

\item{penalty.type}{which method to use}

\item{network.options}{options to be used}
}
\value{

}
\description{
Internal method to calculate the network using data-dependant methods
}
\examples{
xdata <- matrix(rnorm(100), ncol = 20)
calc.penalty(xdata, 'correlation')
calc.penalty(xdata, 'correlation', network.options.default(cutoff = .6))
calc.penalty(xdata, 'covariance')
calc.penalty(xdata, 'covariance', network.options.default(cutoff = .6))
}
