% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{cor.parallel}
\alias{cor.parallel}
\title{Calculate correlation of large matrix}
\usage{
cor.parallel(xdata, method = "pearson", base.dir = network.cox::base.dir(),
  build.matrix = T, n.cores = parallel:::detectCores(),
  force.recalc = FALSE, show.message = FALSE)
}
\arguments{
\item{xdata}{calculate correlation matrix on each column}

\item{build.matrix}{boolean parameter to return the matrix in the end}

\item{n.cores}{number of cores to be used}
}
\value{
a mtarix if
}
\description{
Calculate correlation of large matrix
}
\examples{
n.col <- 6
xdata <- matrix(rnorm(n.col * 4), ncol = n.col)
cor.parallel(xdata)
}
