% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmDegree.R
\name{cv.glmDegree}
\alias{cv.glmDegree}
\title{GLMNET cross-validation model penalizing nodes with small degree}
\usage{
cv.glmDegree(
  xdata,
  ydata,
  network,
  options = networkOptions(),
  experiment = NULL,
  network.options = deprecated(),
  experiment.name = deprecated(),
  ...
)
}
\arguments{
\item{xdata}{input data, can be a matrix or MultiAssayExperiment.}

\item{ydata}{response data compatible with glmnet.}

\item{network}{type of network, see below.}

\item{options}{options to calculate network.}

\item{experiment}{name of experiment to use as input in
MultiAssayExperiment object (only if xdata is an object of this class).}

\item{network.options}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{experiment.name}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{...}{parameters that \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} accepts.}
}
\value{
an object just as \code{cv.glmnet}
}
\description{
This function overrides the \code{trans.fun} options in \code{networkOptions} with the
inverse of a degree described in Veríssimo et al. (2015) that penalizes
nodes with small degree.
}
\examples{
xdata <- matrix(rnorm(100), ncol = 5)
cv.glmDegree(
  xdata,
  rnorm(nrow(xdata)),
  "correlation",
  family = "gaussian",
  nfolds = 5,
  network_options = networkOptions(min.degree = .2)
)
}
\seealso{
Generic function without pre-defined penalization: \code{\link[=cv.glmNetSparse]{cv.glmNetSparse()}}.
Other penalizations: \code{\link[=cv.glmHub]{cv.glmHub()}} and \code{\link[=cv.glmOrphan]{cv.glmOrphan()}}.
Model with the same penalization: \code{\link[=glmDegree]{glmDegree()}}.
}
