% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_contrib.R
\name{cv.glmHub}
\alias{cv.glmHub}
\title{GLMNET cross-validation model penalizing nodes with small degree}
\usage{
cv.glmHub(xdata, ydata, network,
  network.options = network.options.default(), ...)
}
\arguments{
\item{xdata}{input data, can be a matrix or MultiAssayExperiment}

\item{ydata}{response data compatible with glmnet}

\item{network}{type of network, see below}

\item{network.options}{options to calculate network}

\item{...}{parameters that glmnet accepts}
}
\value{
see cv.glmSparseNet
}
\description{
This function overrides the `trans.fun` options in `network.options` with
an heuristic described in Veríssimo et al. that penalizes nodes with small
degree.
}
\examples{
xdata <- matrix(rnorm(100), ncol = 20)
cv.glmHub(xdata, rnorm(nrow(xdata)), 'correlation', family = 'gaussian',
          network.options = network.options.default(min.degree = .2))
}
\seealso{
glmNetSparse
}
