% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.options.R
\name{degree.heuristic}
\alias{degree.heuristic}
\title{Heuristic function to transformate penalty factor}
\usage{
degree.heuristic(x)
}
\arguments{
\item{x}{single value of vector}
}
\value{
transformed
}
\description{
Heuristic function to transformate penalty factor
}
\examples{
degree.heuristic(rnorm(1:10))
}
