% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{degree.sparsebn}
\alias{degree.sparsebn}
\title{Calculate degree of correlation matrix}
\usage{
degree.sparsebn(xdata, lambdas.length, cutoff = 0,
  consider.unweighted = TRUE, n.cores = parallel:::detectCores(),
  show.message = FALSE, force.recalc.degree = FALSE,
  force.recalc.sparsebn = FALSE, ...)
}
\arguments{
\item{xdata}{calculate correlation matrix on each column}

\item{cutoff}{positive value that determines a cutoff value}

\item{consider.unweighted}{consider all edges as 1 if they are greater than 0}

\item{n.cores}{number of cores to be used}

\item{force.recalc}{force recalculation, instead of going to cache}
}
\value{
a vector of the degrees
}
\description{
Calculate degree of correlation matrix
}
\examples{
n.col <- 6
xdata <- matrix(rnorm(n.col * 4), ncol = n.col)
degree.sparsebn(xdata, 50)
}
