% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{degreeCov}
\alias{degreeCov}
\title{Calculate the degree of the covariance network based on xdata}
\usage{
degreeCov(
  xdata,
  cutoff = 0,
  consider.unweighted = FALSE,
  force.recalc.degree = FALSE,
  force.recalc.network = FALSE,
  n.cores = 1,
  ...
)
}
\arguments{
\item{xdata}{calculate correlation matrix on each column}

\item{cutoff}{positive value that determines a cutoff value}

\item{consider.unweighted}{consider all edges as 1 if they are greater than 0}

\item{force.recalc.degree}{force recalculation of penalty weights (but not
the network), instead of going to cache}

\item{force.recalc.network}{force recalculation of network and penalty
weights, instead of going to cache}

\item{n.cores}{number of cores to be used}

\item{...}{extra parameters for cov function}
}
\value{
a vector of the degrees
}
\description{
Calculate the degree of the covariance network based on xdata
}
\examples{
n.col <- 6
xdata <- matrix(rnorm(n.col * 4), ncol = n.col)
degreeCov(xdata)
degreeCov(xdata, cutoff = .5)
degreeCov(xdata, cutoff = .5, consider.unweighted = TRUE)
}
