% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loose_rock_functions.R
\name{digest.cache}
\alias{digest.cache}
\title{Default digest method}
\usage{
digest.cache(val)
}
\arguments{
\item{val}{object to calculate hash over}
}
\value{
a hash of the sha256
}
\description{
Default digest method
}
\examples{
digest.cache(c(1,2,3,4,5))
}
