% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.calculateResult}
\alias{.calculateResult}
\title{Calculate/load result and save if necessary}
\usage{
.calculateResult(path, compression, forceRecalc, showMessage, fun, ...)
}
\arguments{
\item{path}{path to save cache}

\item{compression}{compression used in save}

\item{forceRecalc}{force to recalculate cache}

\item{showMessage}{boolean to show messages}

\item{fun}{function to be called}

\item{...}{arguments to said function
,}
}
\value{
result of fun(...)
}
\description{
This is where the actual work is done
}
\examples{
glmSparseNet:::.calculateResult(
  file.path(tempdir(), "calculate_result.Rdata"),
  "gzip",
  FALSE,
  TRUE,
  sum,
  1, 2, 3
)
}
