% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.createDirectoryForCache}
\alias{.createDirectoryForCache}
\title{Create directories for cache}
\usage{
.createDirectoryForCache(baseDir, parentPath)
}
\arguments{
\item{baseDir}{tentative base dir to create.}

\item{parentPath}{first 4 characters of digest that will become parent
directory for the actual cache file (this reduces number of files per folder)}
}
\value{
a list of updated base.dir and parent.dir
}
\description{
Create directories for cache
}
\examples{
glmSparseNet:::.createDirectoryForCache(tempdir(), "abcd")
\donttest{
glmSparseNet:::.createDirectoryForCache(
  file.path(getwd(), "run-cache"), "abcd"
)
}
}
