% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{.degreeGeneric}
\alias{.degreeGeneric}
\title{Generic function to calculate degree based on data}
\usage{
.degreeGeneric(
  fun = stats::cor,
  fun.prefix = "operator",
  xdata,
  cutoff = 0,
  consider.unweighted = FALSE,
  chunks = 1000,
  force.recalc.degree = FALSE,
  force.recalc.network = FALSE,
  n.cores = 1,
  ...
)
}
\arguments{
\item{fun}{function that will calculate the edge weight between 2 nodes}

\item{fun.prefix}{used to store low-level information on network as it can
become to large to be stored in memory}

\item{xdata}{calculate correlation matrix on each column}

\item{cutoff}{positive value that determines a cutoff value}

\item{consider.unweighted}{consider all edges as 1 if they are greater than 0}

\item{chunks}{calculate function at batches of this value (default is 1000)}

\item{force.recalc.degree}{force recalculation of penalty weights (but not
the network), instead of going to cache}

\item{force.recalc.network}{force recalculation of network and penalty
weights, instead of going to cache}

\item{n.cores}{number of cores to be used}

\item{...}{extra parameters for fun}
}
\value{
a vector of the degrees
}
\description{
The assumption to use this function is that the network represented by a
matrix is symetric and without
any connection the node and itself.
}
