% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{.showMessage}
\alias{.showMessage}
\title{Show messages option in .runCache}
\usage{
.showMessage(showMessage = NULL)
}
\arguments{
\item{showMessage}{boolean indicating to show messages or not}
}
\value{
the show.message option
}
\description{
Show messages option in .runCache
}
\examples{
glmSparseNet:::.showMessage(FALSE)
}
