% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_apis.R
\name{ensemblGeneNames}
\alias{ensemblGeneNames}
\title{Retrieve ensembl gene names from biomaRt}
\usage{
ensemblGeneNames(gene.id, use.cache = TRUE)
}
\arguments{
\item{gene.id}{character vector with gene names}

\item{use.cache}{Boolean indicating if biomaRt cache should be used}
}
\value{
a dataframe with external gene names, ensembl_id
}
\description{
Retrieve ensembl gene names from biomaRt
}
\examples{
\donttest{
    ensemblGeneNames(c('MOB1A','RFLNB', 'SPIC', 'TP53'))
}
}
