% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.options.R
\name{filter.by.experiment}
\alias{filter.by.experiment}
\title{Filter MultiAssayExperiment colData for specific experiment}
\usage{
filter.by.experiment(multi.assay, experiment.name)
}
\arguments{
\item{experiment.name}{}
}
\description{
Filter MultiAssayExperiment colData for specific experiment
}
