% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_contrib.R
\name{glmDegree}
\alias{glmDegree}
\title{GLMNET model penalizing nodes with small degree}
\usage{
glmDegree(xdata, ydata, network, network.options = networkOptions(), ...)
}
\arguments{
\item{xdata}{input data, can be a matrix or MultiAssayExperiment}

\item{ydata}{response data compatible with glmnet}

\item{network}{type of network, see below}

\item{network.options}{options to calculate network}

\item{...}{parameters that glmnet accepts}
}
\value{
see glmNetSparse
}
\description{
This function overrides the `trans.fun` options in `network.options` with the
inverse of a degree described in Veríssimo et al. (2015) that penalizes
nodes with small degree.
}
\examples{
xdata <- matrix(rnorm(100), ncol = 5)
glmDegree(xdata, rnorm(nrow(xdata)), "correlation",
  family = "gaussian",
  network.options = networkOptions(min.degree = .2)
)
}
\seealso{
glmNetSparse
}
