% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_apis.R
\name{hallmarks}
\alias{hallmarks}
\title{Retrieve hallmarks of cancer count for genes}
\usage{
hallmarks(
  genes,
  metric = "count",
  hierarchy = "full",
  generate.plot = TRUE,
  show.message = FALSE
)
}
\arguments{
\item{genes}{gene names}

\item{metric}{see below}

\item{hierarchy}{see below}

\item{generate.plot}{flag to indicate if return object has a ggplot2 object}

\item{show.message}{flag to indicate if run.cache method shows messages}
}
\value{
data.frame with choosen metric and hierarchy
It also returns a vector with genes that do not have any
hallmarks.

See http://chat.lionproject.net/api for more details on the
metric and hallmarks parameters

To standardize the colors in the gradient you can use
scale_fill_gradientn(limits=c(0,1), colours=topo.colors(3)) to
limit between 0 and 1 for cprob and -1 and 1 for npmi
}
\description{
Retrieve hallmarks of cancer count for genes
}
\examples{
\dontrun{
    hallmarks(c('MOB1A', 'RFLNB', 'SPIC'))
    hallmarks(c('MOB1A', 'RFLNB', 'SPIC'), metric = 'cprob')
}
}
