% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{networkCovParallel}
\alias{networkCovParallel}
\title{Calculates the covariance network}
\usage{
networkCovParallel(
  xdata,
  build.output = "matrix",
  n.cores = 1,
  force.recalc.network = FALSE,
  show.message = FALSE,
  ...
)
}
\arguments{
\item{xdata}{base data to calculate network}

\item{build.output}{if output returns a 'matrix', 'vector' of the upper triu
without the diagonal or NULL with any other argument}

\item{n.cores}{number of cores to be used}

\item{force.recalc.network}{force recalculation, instead of going to cache}

\item{show.message}{shows cache operation messages}

\item{...}{extra parameters for fun}
}
\value{
depends on build.output parameter
}
\description{
Calculates the covariance network
}
\examples{
n.col <- 6
xdata <- matrix(rnorm(n.col * 4), ncol = n.col)
networkCovParallel(xdata)
}
