% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.options.R
\name{networkOptions}
\alias{networkOptions}
\title{Setup network options}
\usage{
networkOptions(
  method = "pearson",
  unweighted = TRUE,
  cutoff = 0,
  centrality = "degree",
  min.degree = 0,
  n.cores = 1,
  trans.fun = function(x) {     x }
)
}
\arguments{
\item{method}{in case of correlation and covariance, which method to use}

\item{unweighted}{calculate degree using unweighted network}

\item{cutoff}{cuttoff value in network edges to trim the network}

\item{centrality}{centrality measure to use, currently only supports degree}

\item{min.degree}{minimum value that individual penalty weight can take}

\item{n.cores}{number of cores to use, default to 1

The trans.fun argument takes a function definition that will apply a
transformation to the penalty vector calculated from the degree. This
transformation allows to change how the penalty is applied.}

\item{trans.fun}{see below}
}
\value{
a list of options
}
\description{
Setup network options, such as using weighted or unweighted degree,
 which centrality measure to use
}
\examples{
networkOptions(unweighted = FALSE)
}
\seealso{
glmOrphan glmDegree
}
