% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data_tcga.R
\name{prepare.tcga.survival.data}
\alias{prepare.tcga.survival.data}
\title{Prepare FPKM data from TCGA project (specific tissue)}
\usage{
prepare.tcga.survival.data(project = "brca",
  tissue.type = "primary.solid.tumor", handle.duplicates = "keep_first",
  coding.genes = FALSE)
}
\arguments{
\item{project}{tcga project that has a package avaliable see https://github.com/averissimo/tcga.data}

\item{tissue.type}{type of tissue, can be 'primary.solid.tumor', 'metastatic', etc... depending on project.}

\item{handle.duplicates}{strategy to handle multiple samples for same individual, can take 'keep_first' or 'keep_all'}

\item{coding.genes}{filter the genes to only include coding genes, see loose.rock::coding.genes}
}
\value{
a list with data ready to be used in survival analysis, the 'xdata.raw' and 'ydata.raw' elements
have the full dataset for the specific tissue and the 'xdata' and 'ydata' have been cleaned by handling
patients with multiple samples, removing individuals with event time <= 0, missing and genes that have
standard_deviation == 0. It also returns a sha256 checksum for each of the data
}
\description{
This function will load data and pre-process it to be used in
 survival models.
}
\details{
It will:
 * load data
 * handle duplicate samples for same individal (default is to keep only first)
 * remove individuals with missing vital_status or both follow-up/death time span
 * remove individuals with follow-up/death time span == 0
 * remove genes from RNASeqData (xdata) with standard deviation == 0
}
\examples{
prepare.tcga.survival.data('brca', 'primary.solid.tumor', 'keep_first')
}
