% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_apis.R
\name{protein2EnsemblGeneNames}
\alias{protein2EnsemblGeneNames}
\title{Retrieve ensembl gene ids from proteins}
\usage{
protein2EnsemblGeneNames(ensembl.proteins, use.cache = TRUE)
}
\arguments{
\item{ensembl.proteins}{character vector with gene names in
ensembl_peptide_id format}

\item{use.cache}{Boolean indicating if biomaRt cache should be used}
}
\value{
a dataframe with external gene names, ensembl_peptide_id
}
\description{
Retrieve ensembl gene ids from proteins
}
\examples{
protein2EnsemblGeneNames(c('ENSP00000235382',
                           'ENSP00000233944',
                           'ENSP00000216911'))
}
